+++ yellowstone-grpc-geyser/src/grpc.rs
@@ -12,8 +12,8 @@ use {
             GetVersionRequest, GetVersionResponse, IsBlockhashValidRequest,
             IsBlockhashValidResponse, PingRequest, PongResponse, SubscribeRequest, SubscribeUpdate,
             SubscribeUpdateAccount, SubscribeUpdateAccountInfo, SubscribeUpdateBlock,
-            SubscribeUpdateBlockMeta, SubscribeUpdatePing, SubscribeUpdateSlot,
-            SubscribeUpdateTransaction, SubscribeUpdateTransactionInfo,
+            SubscribeUpdateBlockMeta, SubscribeUpdateEntry, SubscribeUpdatePing,
+            SubscribeUpdateSlot, SubscribeUpdateTransaction, SubscribeUpdateTransactionInfo,
         },
         version::VERSION,
     },
@@ -179,6 +179,15 @@ impl<'a> From<(&'a ReplicaTransactionInfoV2<'a>, u64)> for MessageTransaction {
 }
 
 #[derive(Debug, Clone)]
+pub struct MessageEntry {
+    pub slot: u64,
+    pub index: usize,
+    pub num_hashes: u64,
+    pub hash: Vec<u8>,
+    pub executed_transaction_count: u64,
+}
+
+#[derive(Debug, Clone)]
 pub struct MessageBlock {
     pub parent_slot: u64,
     pub slot: u64,
@@ -244,6 +253,7 @@ pub enum Message {
     Slot(MessageSlot),
     Account(MessageAccount),
     Transaction(MessageTransaction),
+    Entry(MessageEntry),
     Block(MessageBlock),
     BlockMeta(MessageBlockMeta),
 }
@@ -254,6 +264,7 @@ impl Message {
             Self::Slot(msg) => msg.slot,
             Self::Account(msg) => msg.slot,
             Self::Transaction(msg) => msg.slot,
+            Self::Entry(msg) => msg.slot,
             Self::Block(msg) => msg.slot,
             Self::BlockMeta(msg) => msg.slot,
         }
@@ -264,6 +275,7 @@ impl Message {
             Self::Slot(_) => "Slot",
             Self::Account(_) => "Account",
             Self::Transaction(_) => "Transaction",
+            Self::Entry(_) => "Entry",
             Self::Block(_) => "Block",
             Self::BlockMeta(_) => "BlockMeta",
         }
@@ -321,6 +333,7 @@ pub enum MessageRef<'a> {
     Slot(&'a MessageSlot),
     Account(&'a MessageAccount),
     Transaction(&'a MessageTransaction),
+    Entry(&'a MessageEntry),
     Block(MessageBlockRef<'a>),
     BlockMeta(&'a MessageBlockMeta),
 }
@@ -341,6 +354,13 @@ impl<'a> MessageRef<'a> {
             Self::Transaction(message) => UpdateOneof::Transaction(SubscribeUpdateTransaction {
                 transaction: Some(message.transaction.to_proto()),
                 slot: message.slot,
+            }),
+            Self::Entry(message) => UpdateOneof::Entry(SubscribeUpdateEntry {
+                slot: message.slot,
+                index: message.index as u64,
+                num_hashes: message.num_hashes,
+                hash: message.hash.clone(),
+                executed_transaction_count: message.executed_transaction_count,
             }),
             Self::Block(message) => UpdateOneof::Block(SubscribeUpdateBlock {
                 slot: message.slot,
@@ -914,6 +934,7 @@ impl Geyser for GrpcService {
                 transactions: HashMap::new(),
                 blocks: HashMap::new(),
                 blocks_meta: HashMap::new(),
+                entry: HashMap::new(),
                 commitment: None,
                 accounts_data_slice: Vec::new(),
             },